fix:[sc-28] add a common validate helper
