+++ src/routes/auth.route.js
@@ -1,12 +1,13 @@
 const router = require('express').Router();
 const authControllers = require('../controllers/auth.controller');
 const authValidators = require('../validators/auth.validator');
+const userValidators = require('../validators/users.validator');
 const authMiddlewares = require('../middlewares/auth.middleware');
 const commonHelpers = require('../helpers/common.helper');
 
 router.post(
   '/signup',
-  authValidators.validateSignupPayload,
+  userValidators.validateUser,
   authControllers.signup,
   commonHelpers.customResponseHandler
 );
