+++ src/validators/carts.validator.js
@@ -1,5 +1,6 @@
 const Joi = require('joi');
 const commonHelper = require('../helpers/common.helper');
+const validateHelper = require('../helpers/validate.helper');
 
 const validateCartItemSchema = (req, res, next) => {
   try {
@@ -12,21 +13,17 @@ const validateCartItemSchema = (req, res, next) => {
       quantity: Joi.number().required(),
     });
 
-    const { error, value } = schema.validate(req.body);
+    const validateResponse = validateHelper.validateSchemas(schema, req.body);
+    const isValid = validateResponse[0];
+    const value = validateResponse[1];
 
-    if (error) {
-      console.log(error);
-      const errMsg =
-        error.details
-          .map(detail => detail.message)
-          .join(', ')
-          .replaceAll(`"`, '') || 'Cart item validation failed';
-
-      return commonHelper.customErrorHandler(res, errMsg, 422);
+    if (!isValid) {
+      return commonHelper.customErrorHandler(res, value, 422);
     }
 
     req.body = value;
-    next();
+
+    return next();
   } catch (err) {
     console.log('Error validating input fields: ', err);
     return commonHelper.customErrorHandler(res, err.message, 400);
