+++ src/validators/common.validator.js
@@ -1,5 +1,6 @@
 const Joi = require('joi');
 const commonHelper = require('../helpers/common.helper');
+const validateHelper = require('../helpers/validate.helper');
 const constants = require('../constants/constants');
 
 const validateQueryParams = (req, res, next) => {
@@ -19,21 +20,17 @@ const validateQueryParams = (req, res, next) => {
         .optional(),
     });
 
-    const { error, value } = schema.validate(req.query);
+    const validateResponse = validateHelper.validateSchemas(schema, req.body);
+    const isValid = validateResponse[0];
+    const value = validateResponse[1];
 
-    if (error) {
-      console.log('error: ', error);
-      const errMsg =
-        error.details
-          .map(detail => detail.message)
-          .join(', ')
-          .replaceAll(`"`, '') || 'Query schema data validation failed';
-
-      return commonHelper.customErrorHandler(res, errMsg, 422);
+    if (!isValid) {
+      return commonHelper.customErrorHandler(res, value, 422);
     }
 
-    req.query = value; // Update query object with validated values
-    next();
+    req.query = value;
+
+    return next();
   } catch (err) {
     console.log('Error validating dish input fields: ', err);
     return commonHelper.customErrorHandler(res, err.message, 400);
