+++ src/validators/orders.validator.js
@@ -1,5 +1,6 @@
 const Joi = require('joi');
 const commonHelper = require('../helpers/common.helper');
+const validateHelper = require('../helpers/validate.helper');
 const constants = require('../constants/constants');
 
 const validatePlaceOrderSchema = (req, res, next) => {
@@ -17,21 +18,17 @@ const validatePlaceOrderSchema = (req, res, next) => {
         .required(),
     });
 
-    const { error, value } = schema.validate(req.body);
+    const validateResponse = validateHelper.validateSchemas(schema, req.body);
+    const isValid = validateResponse[0];
+    const value = validateResponse[1];
 
-    if (error) {
-      console.log(error);
-      const errMsg =
-        error.details
-          .map(detail => detail.message)
-          .join(', ')
-          .replaceAll(`"`, '') || 'Input fields validation failed for placing order';
-
-      return commonHelper.customErrorHandler(res, errMsg, 422);
+    if (!isValid) {
+      return commonHelper.customErrorHandler(res, value, 422);
     }
 
     req.body = value;
-    next();
+
+    return next();
   } catch (err) {
     console.log('Error validating input fields: ', err);
     return commonHelper.customErrorHandler(res, err.message, 400);
@@ -48,21 +45,17 @@ const validateDeliveryPartnerId = (req, res, next) => {
         .required(),
     });
 
-    const { error, value } = schema.validate(req.body);
+    const validateResponse = validateHelper.validateSchemas(schema, req.body);
+    const isValid = validateResponse[0];
+    const value = validateResponse[1];
 
-    if (error) {
-      console.log(error);
-      const errMsg =
-        error.details
-          .map(detail => detail.message)
-          .join(', ')
-          .replaceAll(`"`, '') || 'Delivery partner id validation failed';
-
-      return commonHelper.customErrorHandler(res, errMsg, 422);
+    if (!isValid) {
+      return commonHelper.customErrorHandler(res, value, 422);
     }
 
     req.body = value;
-    next();
+
+    return next();
   } catch (err) {
     console.log('Error validating input fields: ', err);
     return commonHelper.customErrorHandler(res, err.message, 400);
@@ -77,21 +70,17 @@ const validateOrderStatus = (req, res, next) => {
         .valid(...Object.values(constants.ORDER_STATUS)),
     });
 
-    const { error, value } = schema.validate(req.body);
+    const validateResponse = validateHelper.validateSchemas(schema, req.body);
+    const isValid = validateResponse[0];
+    const value = validateResponse[1];
 
-    if (error) {
-      console.log(error);
-      const errMsg =
-        error.details
-          .map(detail => detail.message)
-          .join(', ')
-          .replaceAll(`"`, '') || 'Delivery partner id validation failed';
-
-      return commonHelper.customErrorHandler(res, errMsg, 422);
+    if (!isValid) {
+      return commonHelper.customErrorHandler(res, value, 422);
     }
 
     req.body = value;
-    next();
+
+    return next();
   } catch (err) {
     console.log('Error validating input fields: ', err);
     return commonHelper.customErrorHandler(res, err.message, 400);
