+++ src/validators/users.validator.js
@@ -1,29 +1,57 @@
 const Joi = require('joi');
 const commonHelper = require('../helpers/common.helper');
+const validateHelper = require('../helpers/validate.helper');
 
+const validateUser = (req, res, next) => {
+  try {
+    const schema = Joi.object({
+      first_name: Joi.string().required().min(3),
+      last_name: Joi.string().required().min(3),
+      email: Joi.string()
+        .required()
+        .email({
+          minDomainSegments: 2,
+          tlds: { allow: ['com', 'co', 'net'] },
+        }),
+      phone: Joi.string().required().min(10).max(15),
+      address: Joi.string().min(5),
+    });
+
+    const validateResponse = validateHelper.validateSchemas(schema, req.body);
+    const isValid = validateResponse[0];
+    const value = validateResponse[1];
+
+    if (!isValid) {
+      return commonHelper.customErrorHandler(res, value, 422);
+    }
+
+    req.body = value;
+    return next();
+  } catch (err) {
+    console.log('Error validating input fields: ', err);
+    return commonHelper.customErrorHandler(res, err.message, 400);
+  }
+};
 const validateUserAddress = (req, res, next) => {
   try {
     const schema = Joi.object({
       address: Joi.string().required(),
     });
 
-    const { error, value } = schema.validate(req.body);
-    if (error) {
-      console.log('error: ', error);
-      const errMsg =
-        error.details
-          .map(detail => detail.message)
-          .join(', ')
-          .replaceAll(`"`, '') || 'User address failed';
-      return commonHelper.customErrorHandler(res, errMsg, 400);
+    const validateResponse = validateHelper.validateSchemas(schema, req.body);
+    const isValid = validateResponse[0];
+    const value = validateResponse[1];
+
+    if (!isValid) {
+      return commonHelper.customErrorHandler(res, value, 422);
     }
 
     req.body = value;
-    next();
+    return next();
   } catch (err) {
     console.log('Error validating input fields: ', err);
     return commonHelper.customErrorHandler(res, err.message, 400);
   }
 };
 
-module.exports = { validateUserAddress };
+module.exports = { validateUser, validateUserAddress };
